screener_app/requirements.txt
flask
numpy 
pandas
gunicorn
praw
yfinance
matplotlib
nltk
google-cloud-storage